# src/main/java/com/mmsp/dto/DeviceRegisterRequest.java
package com.mmsp.dto;

import jakarta.validation.constraints.NotBlank;

import java.util.Map;

public class DeviceRegisterRequest {

    @NotBlank
    private String deviceSerial;

    @NotBlank
    private String deviceType;

    private String firmwareVersion;
    private String siteId;
    private Map<String, Object> metadata;

    public String getDeviceSerial() {
        return deviceSerial;
    }

    public void setDeviceSerial(String deviceSerial) {
        this.deviceSerial = deviceSerial;
    }

    public String getDeviceType() {
        return deviceType;
    }

    public void setDeviceType(String deviceType) {
        this.deviceType = deviceType;
    }

    public String getFirmwareVersion() {
        return firmwareVersion;
    }

    public void setFirmwareVersion(String firmwareVersion) {
        this.firmwareVersion = firmwareVersion;
    }

    public String getSiteId() {
        return siteId;
    }

    public void setSiteId(String siteId) {
        this.siteId = siteId;
    }

    public Map<String, Object> getMetadata() {
        return metadata;
    }

    public void setMetadata(Map<String, Object> metadata) {
        this.metadata = metadata;
    }
}