# src/main/java/com/mmsp/dto/HeartbeatRequest.java
package com.mmsp.dto;

public class HeartbeatRequest {
    private String timestamp;
    private Double cpuPercent;
    private Double memoryPercent;

    public String getTimestamp() {
        return timestamp;
    }

    public void setTimestamp(String timestamp) {
        this.timestamp = timestamp;
    }

    public Double getCpuPercent() {
        return cpuPercent;
    }

    public void setCpuPercent(Double cpuPercent) {
        this.cpuPercent = cpuPercent;
    }

    public Double getMemoryPercent() {
        return memoryPercent;
    }

    public void setMemoryPercent(Double memoryPercent) {
        this.memoryPercent = memoryPercent;
    }
}